requirements.txt
streamlit
gymnasium
Pillow
